[core] implement function to compile virtual python module from generated/templated code
